+++ openshift-client/src/main/java/io/fabric8/openshift/client/dsl/OpenShiftClusterAutoscalingAPIGroupDSL.java
@@ -13,7 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.knative.client.serving.v1beta1;
+package io.fabric8.openshift.client.dsl;
 
-public interface NamespacedServingV1Beta1Client extends ServingV1Beta1Client, GenericServingV1Beta1Client<NamespacedServingV1Beta1Client> {
+import io.fabric8.kubernetes.client.Client;
+
+public interface OpenShiftClusterAutoscalingAPIGroupDSL extends Client {
+  V1ClusterAutoscalingAPIGroupDSL v1();
+  V1beta1ClusterAutoscalingAPIGroupDSL v1beta1();
 }