+++ openshift-client/src/main/java/io/fabric8/openshift/client/dsl/internal/OpenShiftCreateOnlyResourceNonNamespaceOperationsImpl.java
@@ -13,12 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.openshift.client;
+package io.fabric8.openshift.client.dsl.internal;
 
 import io.fabric8.kubernetes.client.Config;
 import io.fabric8.kubernetes.client.KubernetesClientException;
 import io.fabric8.kubernetes.client.dsl.InOutCreateable;
-import io.fabric8.kubernetes.client.dsl.Namespaceable;
 import io.fabric8.kubernetes.client.dsl.base.OperationContext;
 import io.fabric8.kubernetes.client.dsl.base.OperationSupport;
 import okhttp3.OkHttpClient;
@@ -26,37 +25,33 @@ import okhttp3.OkHttpClient;
 import java.io.IOException;
 import java.util.concurrent.ExecutionException;
 
-public class OpenShiftLocalSubjectAccessReviewOperationsImpl<I, O> extends OperationSupport implements InOutCreateable<I, O>, Namespaceable<OpenShiftLocalSubjectAccessReviewOperationsImpl<I, O>> {
-  private final String subjectAccessApiGroupName;
-  private final String subjectAccessApiGroupVersion;
-  private final String plural;
-  private final Class<O> responseClass;
 
-  public OpenShiftLocalSubjectAccessReviewOperationsImpl(OkHttpClient client, Config config, String apiGroupName, String apiGroupVersion, String plural, Class<O> responseClass) {
-    this(new OperationContext().withOkhttpClient(client).withConfig(config), apiGroupName, apiGroupVersion, plural, responseClass);
+public class OpenShiftCreateOnlyResourceNonNamespaceOperationsImpl<I, O> extends OperationSupport implements InOutCreateable<I, O> {
+  private final Class<O> responseType;
+
+  public OpenShiftCreateOnlyResourceNonNamespaceOperationsImpl(OkHttpClient client, Config config, String apiGroupName, String apiGroupVersion, String plural, Class<O> responseType) {
+    this(new OperationContext().withOkhttpClient(client).withConfig(config), apiGroupName, apiGroupVersion, plural, responseType);
   }
 
-  public OpenShiftLocalSubjectAccessReviewOperationsImpl(OperationContext context, String apiGroupName, String apiGroupVersion, String plural, Class<O> responseClass) {
+  public OpenShiftCreateOnlyResourceNonNamespaceOperationsImpl(OperationContext context, String apiGroupName, String apiGroupVersion, String plural, Class<O> responseType) {
     super(context.withApiGroupName(apiGroupName)
       .withApiGroupVersion(apiGroupVersion)
       .withPlural(plural));
-    this.subjectAccessApiGroupName = apiGroupName;
-    this.subjectAccessApiGroupVersion = apiGroupVersion;
-    this.plural = plural;
-    this.responseClass = responseClass;
+    this.responseType = responseType;
   }
 
+  @SafeVarargs
   @Override
-  public O create(I... resources) {
+  public final O create(I... resources) {
     try {
       if (resources.length > 1) {
         throw new IllegalArgumentException("Too many items to create.");
       } else if (resources.length == 1) {
-        return handleCreate(resources[0], responseClass);
+        return handleCreate(resources[0], responseType);
       } else if (getItem() == null) {
         throw new IllegalArgumentException("Nothing to create.");
       } else {
-        return handleCreate(getItem(), responseClass);
+        return handleCreate(getItem(), responseType);
       }
     } catch (ExecutionException | IOException e) {
       throw KubernetesClientException.launderThrowable(e);
@@ -69,7 +64,7 @@ public class OpenShiftLocalSubjectAccessReviewOperationsImpl<I, O> extends Opera
   @Override
   public O create(I item) {
     try {
-      return handleCreate(item, responseClass);
+      return handleCreate(item, responseType);
     } catch (ExecutionException | IOException e) {
       throw KubernetesClientException.launderThrowable(e);
     } catch (InterruptedException ie) {
@@ -79,13 +74,11 @@ public class OpenShiftLocalSubjectAccessReviewOperationsImpl<I, O> extends Opera
   }
 
   @Override
-  public OpenShiftLocalSubjectAccessReviewOperationsImpl<I, O> inNamespace(String namespace) {
-    this.namespace = namespace;
-    return new OpenShiftLocalSubjectAccessReviewOperationsImpl<>(context.withNamespace(namespace), subjectAccessApiGroupName, subjectAccessApiGroupVersion, this.plural, responseClass);
+  public boolean isResourceNamespaced() {
+    return false;
   }
 
   public I getItem() {
     return (I) context.getItem();
   }
 }
-